requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
cffi                1.14.2
cryptography        3.1
Django              3.1.1
django-rest-knox    4.1.0
django-solo         1.1.3
djangorestframework 3.11.1
Pillow              7.2.0
pip                 20.2.3
pycparser           2.20
python-decouple     3.3
pytz                2020.1
setuptools          49.6.0
six                 1.15.0
sqlparse            0.3.1
wheel               0.35.1
